# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
...
        "/api/webinarLectors": {
            "get": {
                "tags": [
                    "WEBINARS"
                ],
                "summary": "Получение массива лекторов для выпадающего списка фильтрации",
                "description": "Получение массива лекторов для выпадающего списка фильтрации",
                "operationId": "getWebinarLectors",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}